pyproject.toml
[tool.poetry]
name = "venv"
version = "0.1.0"
description = "FastAPI Structure"
authors = ["TheHormat"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.35"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
alembic = "^1.13.3"
google-generativeai = "^0.8.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
